# docs/swagger.yaml
basePath: /api
definitions:
...
  entity.EducationLevelEnum:
    enum:
    - 1 - Doctoral / Professor
    - 2 - Master Degree
    - 3 - Bachelor
    - 4 - Diploma 1
    - 5 - Diploma 2
    - 6 - Diploma 3
    - 7 - Diploma 4
    - 8 - Elementary School
    - 9 - Senior High School
    - 10 - Junior High School
    - 11 - Unschooled
    type: string
    x-enum-varnames:
    - EDUCATION_LEVEL_ENUM_DOCTORAL
    - EDUCATION_LEVEL_ENUM_MASTER
    - EDUCATION_LEVEL_ENUM_BACHELOR
    - EDUCATION_LEVEL_ENUM_D1
    - EDUCATION_LEVEL_ENUM_D2
    - EDUCATION_LEVEL_ENUM_D3
    - EDUCATION_LEVEL_ENUM_D4
    - EDUCATION_LEVEL_ENUM_SD
    - EDUCATION_LEVEL_ENUM_SMA
    - EDUCATION_LEVEL_ENUM_SMP
    - EDUCATION_LEVEL_ENUM_TK
...
  entity.MaritalStatusEnum:
    enum:
    - single
    - married
    - divorced
    - widowed
    - any
    type: string
    x-enum-varnames:
    - MARITAL_STATUS_ENUM_SINGLE
    - MARITAL_STATUS_ENUM_MARRIED
    - MARITAL_STATUS_ENUM_DIVORCED
    - MARITAL_STATUS_ENUM_WIDOWED
    - MARITAL_STATUS_ENUM_ANY
...
  entity.UserGender:
    enum:
    - MALE
    - FEMALE
    type: string
    x-enum-varnames:
    - MALE
    - FEMALE
...
  entity.UserStatus:
    enum:
    - ACTIVE
    - INACTIVE
    - PENDING
    type: string
    x-enum-varnames:
    - USER_ACTIVE
    - USER_INACTIVE
    - USER_PENDING
...
  request.UpdateStatusUserProfileRequest:
    properties:
      id:
        type: string
      status:
        type: string
    required:
    - id
    - status
    type: object
...
  response.EducationResponse:
    properties:
      certificate:
        type: string
      created_at:
        type: string
      education_level:
        $ref: '#/definitions/entity.EducationLevelEnum'
      end_date:
        type: string
      gpa:
        type: number
      graduate_year:
        type: integer
      id:
        type: string
      major:
        type: string
      school_name:
        type: string
      updated_at:
        type: string
      user_profile_id:
        type: string
    type: object
...
  response.SkillResponse:
    properties:
      certificate:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      level:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_profile_id:
        type: string
    type: object
...
  response.UserProfileResponse:
    properties:
      age:
        type: integer
      birth_date:
        type: string
      birth_place:
        type: string
      created_at:
        type: string
      curriculum_vitae:
        type: string
      educations:
        items:
          $ref: '#/definitions/response.EducationResponse'
        type: array
      gender:
        $ref: '#/definitions/entity.UserGender'
      id:
        type: string
      ktp:
        type: string
      marital_status:
        $ref: '#/definitions/entity.MaritalStatusEnum'
      name:
        type: string
      phone_number:
        type: string
      skills:
        items:
          $ref: '#/definitions/response.SkillResponse'
        type: array
      status:
        $ref: '#/definitions/entity.UserStatus'
      updated_at:
        type: string
      user:
        additionalProperties: true
        type: object
      user_id:
        type: string
      work_experiences:
        items:
          $ref: '#/definitions/response.WorkExperienceResponse'
        type: array
    type: object
  response.WorkExperienceResponse:
    properties:
      certificate:
        type: string
      company_name:
        type: string
      created_at:
        type: string
      id:
        type: string
      job_description:
        type: string
      name:
        type: string
      updated_at:
        type: string
      user_profile_id:
        type: string
      year_experience:
        type: integer
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Julong Recruitment API Docs
  version: "1.0"
paths:
...
  /user-profiles:
    get:
      consumes:
      - application/json
      description: Find all user profiles paginated
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: page_size
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      - description: Created At
        in: query
        name: created_at
        type: string
      - description: Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserProfileResponse'
      security:
      - BearerAuth: []
      summary: Find all user profiles paginated
      tags:
      - User Profiles
    post:
      consumes:
      - multipart/form-data
      description: Fill user profile
      parameters:
      - description: ID
        in: formData
        name: id
        type: string
      - description: Name
        in: formData
        name: name
        type: string
      - description: Marital Status
        in: formData
        name: marital_status
        required: true
        type: string
      - description: Gender
        in: formData
        name: gender
        required: true
        type: string
      - description: Phone Number
        in: formData
        name: phone_number
        required: true
        type: string
      - description: Age
        in: formData
        name: age
        required: true
        type: integer
      - description: Birth Date
        in: formData
        name: birth_date
        required: true
        type: string
      - description: Birth Place
        in: formData
        name: birth_place
        required: true
        type: string
      - description: KTP
        in: formData
        name: ktp
        type: file
      - description: Curriculum Vitae
        in: formData
        name: curriculum_vitae
        type: file
      - description: KTP Path
        in: formData
        name: ktp_path
        type: string
      - description: CV Path
        in: formData
        name: cv_path
        type: string
      - description: Work Experience ID
        in: formData
        name: work_experiences.id
        type: string
      - description: Work Experience Name
        in: formData
        name: work_experiences.name
        type: string
      - description: Work Experience Company Name
        in: formData
        name: work_experiences.company_name
        type: string
      - description: Work Experience Year
        in: formData
        name: work_experiences.year_experience
        type: integer
      - description: Work Experience Job Description
        in: formData
        name: work_experiences.job_description
        type: string
      - description: Work Experience Certificate
        in: formData
        name: work_experiences.certificate
        type: file
      - description: Education ID
        in: formData
        name: educations.id
        type: string
      - description: Education Level
        in: formData
        name: educations.education_level
        type: string
      - description: Education Major
        in: formData
        name: educations.major
        type: string
      - description: Education School Name
        in: formData
        name: educations.school_name
        type: string
      - description: Education Graduate Year
        in: formData
        name: educations.graduate_year
        type: integer
      - description: Education End Date
        in: formData
        name: educations.end_date
        type: string
      - description: Education Certificate
        in: formData
        name: educations.certificate
        type: file
      - description: Education GPA
        in: formData
        name: educations.gpa
        type: number
      - description: Skill ID
        in: formData
        name: skills.id
        type: string
      - description: Skill Name
        in: formData
        name: skills.name
        type: string
      - description: Skill Description
        in: formData
        name: skills.description
        type: string
      - description: Skill Level
        in: formData
        name: skills.level
        type: integer
      - description: Skill Certificate
        in: formData
        name: skills.certificate
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserProfileResponse'
      security:
      - BearerAuth: []
      summary: Fill user profile
      tags:
      - User Profiles
  /user-profiles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user profile
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete user profile
      tags:
      - User Profiles
    get:
      consumes:
      - application/json
      description: Find user profile by id
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserProfileResponse'
      security:
      - BearerAuth: []
      summary: Find user profile by id
      tags:
      - User Profiles
  /user-profiles/update-status:
    put:
      consumes:
      - application/json
      description: Update status user profile
      parameters:
      - description: Update Status User Profile
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateStatusUserProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserProfileResponse'
      security:
      - BearerAuth: []
      summary: Update status user profile
      tags:
      - User Profiles
  /user-profiles/user:
    get:
      consumes:
      - application/json
      description: Find user profile by user id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserProfileResponse'
      security:
      - BearerAuth: []
      summary: Find user profile by user id
      tags:
      - User Profiles
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"